+++ pages/api/admin/keepandsend/keepone/index.js
@@ -1,23 +1,38 @@
-import {postKeepByTypeNumberLottoDateId, putKeep, deleteKeep} from '../../../../../database/controller/keepandsend'
+import {
+  postKeepByTypeNumberLottoDateId,
+  putKeep,
+  deleteKeep,
+} from "../../../../../database/controller/keepandsend";
+import { getToken } from "next-auth/jwt";
+import {
+  responseError,
+  responseSuccess,
+} from "../../../../../lib/responseJson";
 
-export default async function handler(req, res){
-  switch(req.method){
-    case 'POST': 
-      postKeepByTypeNumberLottoDateId(req, res)
-      break
-    case 'PUT': 
-      putKeep(req, res)
-      break
-    // case 'GET': 
-    //   getKeepAllByTypeAndLottoDateId(req, res)
-    //   break
-    case 'DELETE': 
-      deleteKeep(req, res)
-      break
-    default:
-      res.setHeader('Allow', ['POST', 'PUT', 'DELETE'])
-      // res.setHeader('Allow', ['GET', 'POST', 'PUT', 'DELETE'])
-      res.status(405).end(`Method ${req.method} Not Allowed`)
-      break
+export default async function handler(req, res) {
+  const token = await getToken({ req });
+  if (token.role === "admin") {
+    switch (req.method) {
+      case "POST":
+        postKeepByTypeNumberLottoDateId(req, res);
+        break;
+      case "PUT":
+        putKeep(req, res);
+        break;
+      case "DELETE":
+        deleteKeep(req, res);
+        break;
+      default:
+        res.setHeader("Allow", ["POST", "PUT", "DELETE"]);
+        // res.setHeader('Allow', ['GET', 'POST', 'PUT', 'DELETE'])
+        res.status(405).end(`Method ${req.method} Not Allowed`);
+        break;
+    }
+  } else {
+    responseError(
+      res,
+      403,
+      "getKeepAndSendDataByTypeAndLottoDateId protect api by token"
+    );
   }
 }
